Remove py26 support from bindep
as of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.

Change-Id: I6dce75d0d1e015e690fe2b7c5e739c10bc12d48e
Related-Bug: #1519510