bindep: depends: Add support for openSUSE {Leap,Tumbleweed}
lsb_release -cirs reports the following for openSUSE Leap and
Tumbleweed respectively

- openSUSE Tumbleweed 20170213 n/a
- openSUSE project 42.2 n/a

As a result of which, 'bindep --profiles' prints the following

- platform:opensusetumbleweed
[...]
- platform:opensuseproject
[...]

So we need to assign the 'rpm' class to these openSUSE distributions
as well. We also make the 'opensuse' platform an alias to these
two openSUSE distributions since no actively maintained openSUSE
distribution returns just 'openSUSE' in lsb_release -cirs.

Change-Id: Id7104045c7f126f70332322a5e4dbf2f5921096c