bindep: Query for RPM capabilities as well
It's possible for a bindep file to list capabilities instead of actual
package names so it can cover different packages that provide the same
functionality. It's also possible to use capabilities to 'capture'
multiple releases of the same distribution that may have different
rpm package names for the same application. RPMs normally provide at
least one capability named after their actual package name so this
change actual extends the behavior of bindep without alterning it.

For example, here are some results on the openSUSE Leap 42.2
distribution where the package name is 'mariadb'

> rpm -q mariadb
mariadb-10.0.30-20.4.1.x86_64
> rpm --whatprovides -q mariadb
mariadb-10.0.30-20.4.1.x86_64
> rpm -q mariadb-server
package mariadb-server is not installed
> rpm --whatprovides -q mariadb-server
mariadb-10.0.30-20.4.1.x86_64

As we can see, using '--whatprovides' results in bindep picking up
the correct package name based on the capabilities of that package.

Change-Id: I36945d3528013250f3a8617ea63ba1883733f91c