Use the distro package to determine distro
Some distros don't have lsb_release and some aren't even linuxes. Lets
work to support them by using the python distro package which has a
consistent api to getting the distro id, codename, and version info that
we need.

This does its best to be backward compatible by adding atoms for the
os-release reported name and the lsb_release reported name when they
differ on distros; however, it is likely that my survey was incomplete
and there may be other non matching distros. (For some reason it seems
like this is super common with rpm based distros...).

Note that the test for a missing lsb_release has been removed since that
is not the distro package's responsibility and it has multiple
fallbacks.

Change-Id: I891f5d21accb3fbf6d56bd6b0c2ebc2601682442