Fix user profile display
Without this patch, issuing 'bindep --profiles' would fail
on groups that aren't a 2-tuple, because we assume
always two elements and unpack these two [1].

This is a problem, because in the case of loci [2], it prevents
a user from predicting the list of packages to be installed
on the system in a read only fashion.

Additionally, the current --profile function is showing
platform details under "Configuration profiles" AND
"Platform Profiles", as platforms are not filtered by the
profiles() function.

This patch solves the two linked problems by:
- ensuring the lists of profiles are flattened
- ensure the profiles are following the _partition function.

[1]: https://github.com/openstack-infra/bindep/blob/cec3a7576bc7599f97a705a42e3572133b4891a1/bindep/depends.py#L276
[2]: https://github.com/openstack/loci/blob/8982c3ae717578eb8e565fd75e78a800970dbb25/bindep.txt#L13

Change-Id: Ied4ac178e20daf35f105e2acae0a2b3e11219a56