Fix bindep path
Project config's site variables is already overriding the
bindep_command variable [1], which is used in the bindep
role. It's therefore impossible to override it with
the path given in the playbook.

This patch doesn't fully need the bindep role, as the
whole purpose of this repo is to test bindep's behavior.

Therefore this changes the play to use the checked-in
bindep to install the appropriate packages, instead
of /usr/bindep-env/bin/bindep.

[1]: http://git.openstack.org/cgit/openstack-infra/project-config/tree/zuul/site-variables.yaml#n9

Change-Id: Iff71edaf20ef81c16eebfd3fe57c4e809d007e89